+++ modql-macros/src/derives_rusqlite/mod.rs
@@ -1,7 +1,7 @@
-mod from_sqlite_row;
-mod from_sqlite_value;
-mod to_sqlite_value;
+mod sqlite_from_row;
+mod sqlite_from_value;
+mod sqlite_to_value;
 
-pub(crate) use from_sqlite_row::*;
-pub(crate) use from_sqlite_value::*;
-pub(crate) use to_sqlite_value::*;
+pub(crate) use sqlite_from_row::*;
+pub(crate) use sqlite_from_value::*;
+pub(crate) use sqlite_to_value::*;