+++ src/field/sea/has_sea_fields.rs
@@ -11,7 +11,7 @@ pub trait HasSeaFields {
 	/// Return the sea_query::DynIden for each field (just matching the field name)
 	fn sea_idens() -> Vec<DynIden>;
 
-	/// Returns the list of column refs (takes the eventual #[modql(table = "table_name")])
+	/// Returns the list of column refs (takes the eventual #[field(rel = "table_name")])
 	fn sea_column_refs() -> Vec<ColumnRef>;
 
 	/// Returns the list of column refs with the given relation (e.g., table name) and IntoIden (.e.g., StringIden or SIden)